refactor:create utils folder and cartSliceUtils.js
