+++ frontend/src/slices/cartSlice.js
@@ -1,10 +1,9 @@
 import { createSlice } from '@reduxjs/toolkit'
+import { updateCart } from '../utils/cartSliceUtils';
 
 const initialState = localStorage.getItem('cart') ? JSON.parse(localStorage.getItem('cart')) : { cartItems: [] }
 
-const addDecimals = (num) => {
-    return (Math.round(num * 100) / 100).toFixed(2)
-}
+
 
 const cartSlice = createSlice({
     name: 'cart',
@@ -19,22 +18,8 @@ const cartSlice = createSlice({
             } else {
                 state.cartItems = [...state.cartItems, newItem]
             }
-            state.cartItems.map((item) => console.log(item.price))
-            //Calculate items price
-            state.itemsPrice = addDecimals(state.cartItems.reduce((acc, item) => acc + item.price * item.qty,0))
-            //Calculate shipping price
-            state.shippingPrice = addDecimals(state.itemsPrice > 100 ? 0 : 10);
-            //Calculate tax price
-            state.taxPrice = addDecimals(Number((0.15 * state.itemsPrice).toFixed(2)))
-            //Calculate total price
-            state.totalPrice = (
-                Number(state.itemsPrice) +
-                Number(state.shippingPrice) +
-                Number(state.taxPrice)
-            ).toFixed(2)
 
-            localStorage.setItem('cart', JSON.stringify(state))
-
+            return updateCart(state)
         }
     }
 })
